+++ hutool-aop/src/main/java/cn/hutool/aop/aspects/TimeIntervalAspect.java
@@ -3,8 +3,7 @@ package cn.hutool.aop.aspects;
 import java.lang.reflect.Method;
 
 import cn.hutool.core.date.TimeInterval;
-import cn.hutool.log.Log;
-import cn.hutool.log.LogFactory;
+import cn.hutool.core.lang.Console;
 
 /**
  * 通过日志打印方法的执行时间的切面
@@ -12,12 +11,7 @@ import cn.hutool.log.LogFactory;
  *
  */
 public class TimeIntervalAspect extends SimpleAspect{
-	private static final Log log = LogFactory.get();
 
-	public TimeIntervalAspect(Object target) {
-		super(target);
-	}
-	
 	private TimeInterval interval = new TimeInterval();
 
 	@Override
@@ -28,7 +22,7 @@ public class TimeIntervalAspect extends SimpleAspect{
 	
 	@Override
 	public boolean after(Object target, Method method, Object[] args) {
-		log.info("Method [{}.{}] execute spend [{}]ms", target.getClass().getName(), method.getName(), interval.intervalMs());
+		Console.log("Method [{}.{}] execute spend [{}]ms", target.getClass().getName(), method.getName(), interval.intervalMs());
 		return true;
 	}
 }