+++ hutool-core/src/main/java/cn/hutool/core/io/IoUtil.java
@@ -696,7 +696,21 @@ public class IoUtil {
 		if (content == null) {
 			return null;
 		}
-		return new ByteArrayInputStream(StrUtil.bytes(content, charset));
+		return toStream(StrUtil.bytes(content, charset));
+	}
+	
+	/**
+	 * String 转为流
+	 * 
+	 * @param content 内容bytes
+	 * @return 字节流
+	 * @since 4.1.8
+	 */
+	public static ByteArrayInputStream toStream(byte[] content) {
+		if (content == null) {
+			return null;
+		}
+		return new ByteArrayInputStream(content);
 	}
 
 	/**