+++ hutool-core/src/main/java/cn/hutool/core/lang/Validator.java
@@ -428,6 +428,31 @@ public class Validator {
 			throw new ValidateException(errorMsg);
 		}
 	}
+	
+	/**
+	 * 验证该字符串是否是字母（包括大写和小写字母）
+	 * 
+	 * @param value 字符串内容
+	 * @return 是否是字母（包括大写和小写字母）
+	 * @since 4.1.8
+	 */
+	public static boolean isWord(String value) {
+		return isMactchRegex(PatternPool.WORD, value);
+	}
+	
+	/**
+	 * 验证是否为字母（包括大写和小写字母）
+	 * 
+	 * @param value 表单值
+	 * @param errorMsg 验证错误的信息
+	 * @throws ValidateException 验证异常
+	 * @since 4.1.8
+	 */
+	public static void validateWord(String value, String errorMsg) throws ValidateException {
+		if (false == isWord(value)) {
+			throw new ValidateException(errorMsg);
+		}
+	}
 
 	/**
 	 * 验证是否为货币
@@ -673,7 +698,7 @@ public class Validator {
 	 * 验证是否为中国车牌号
 	 * 
 	 * @param value 值
-	 * @return 是否为IPV4地址
+	 * @return 是否为中国车牌号
 	 * @since 3.0.6
 	 */
 	public static boolean isPlateNumber(String value) {