+++ hutool-core/src/test/java/cn/hutool/core/lang/ClassScanerTest.java
@@ -2,15 +2,14 @@ package cn.hutool.core.lang;
 
 import java.util.Set;
 
-import org.junit.Ignore;
 import org.junit.Test;
 
 public class ClassScanerTest {
 	
 	@Test
-	@Ignore
+//	@Ignore
 	public void scanTest() {
-		ClassScaner scaner = new ClassScaner("cn.hutool.core", null);
+		ClassScaner scaner = new ClassScaner("cn.hutool.core.util.StrUtil", null);
 		Set<Class<?>> set = scaner.scan();
 		for (Class<?> clazz : set) {
 			Console.log(clazz.getName());