+++ hutool-http/src/main/java/cn/hutool/http/webservice/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Webservice客户端封装实现
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.http.webservice;